+++ hadoop-hdds/framework/src/main/java/org/apache/hadoop/hdds/security/x509/keys/SecurityUtil.java
@@ -34,66 +34,18 @@ import org.apache.hadoop.hdds.security.ssl.PemFileBasedKeyStoresFactory;
 import org.apache.hadoop.hdds.security.x509.SecurityConfig;
 import org.apache.hadoop.hdds.security.x509.certificate.client.CertificateClient;
 import org.apache.hadoop.hdds.security.x509.exception.CertificateException;
-import org.bouncycastle.asn1.ASN1ObjectIdentifier;
-import org.bouncycastle.asn1.ASN1Sequence;
-import org.bouncycastle.asn1.ASN1Set;
-import org.bouncycastle.asn1.pkcs.Attribute;
-import org.bouncycastle.asn1.pkcs.PKCSObjectIdentifiers;
-import org.bouncycastle.asn1.x500.X500Name;
-import org.bouncycastle.asn1.x509.Extensions;
-import org.bouncycastle.pkcs.PKCS10CertificationRequest;
 
 /**
  * Utility functions for Security modules for Ozone.
  */
 public final class SecurityUtil {
 
-  // Ozone Certificate distinguished format: (CN=Subject,OU=ScmID,O=ClusterID).
-  private static final String DISTINGUISHED_NAME_FORMAT = "CN=%s,OU=%s,O=%s";
-
   private SecurityUtil() {
   }
 
-  public static String getDistinguishedNameFormat() {
-    return DISTINGUISHED_NAME_FORMAT;
-  }
-
-  public static X500Name getDistinguishedName(String subject, String scmID,
-      String clusterID) {
-    return new X500Name(String.format(getDistinguishedNameFormat(), subject,
-        scmID, clusterID));
-  }
-
-  // TODO: move the PKCS10CSRValidator class
-  public static Extensions getPkcs9Extensions(PKCS10CertificationRequest csr)
-      throws CertificateException {
-    ASN1Set pkcs9ExtReq = getPkcs9ExtRequest(csr);
-    Object extReqElement = pkcs9ExtReq.getObjects().nextElement();
-    if (extReqElement instanceof Extensions) {
-      return (Extensions) extReqElement;
-    } else {
-      if (extReqElement instanceof ASN1Sequence) {
-        return Extensions.getInstance((ASN1Sequence) extReqElement);
-      } else {
-        throw new CertificateException("Unknown element type :" + extReqElement
-            .getClass().getSimpleName());
-      }
-    }
-  }
-
-  public static ASN1Set getPkcs9ExtRequest(PKCS10CertificationRequest csr)
-      throws CertificateException {
-    for (Attribute attr : csr.getAttributes()) {
-      ASN1ObjectIdentifier oid = attr.getAttrType();
-      if (oid.equals(PKCSObjectIdentifiers.pkcs_9_at_extensionRequest)) {
-        return attr.getAttrValues();
-      }
-    }
-    throw new CertificateException("No PKCS#9 extension found in CSR");
-  }
-
-  /*
+  /**
    * Returns private key created from encoded key.
+   *
    * @return private key if successful else returns null.
    */
   public static PrivateKey getPrivateKey(byte[] encodedKey,
@@ -117,8 +69,9 @@ public final class SecurityUtil {
     return pvtKey;
   }
 
-  /*
+  /**
    * Returns public key created from encoded key.
+   *
    * @return public key if successful else returns null.
    */
   public static PublicKey getPublicKey(byte[] encodedKey,
