+++ hadoop-hdds/framework/src/test/java/org/apache/hadoop/hdds/security/x509/certificate/utils/TestCertificateSignRequest.java
@@ -22,7 +22,6 @@ import org.apache.hadoop.hdds.conf.OzoneConfiguration;
 import org.apache.hadoop.hdds.security.exception.SCMSecurityException;
 import org.apache.hadoop.hdds.security.x509.SecurityConfig;
 import org.apache.hadoop.hdds.security.x509.keys.HDDSKeyGenerator;
-import org.apache.hadoop.hdds.security.x509.keys.SecurityUtil;
 import org.bouncycastle.asn1.ASN1Encodable;
 import org.bouncycastle.asn1.ASN1ObjectIdentifier;
 import org.bouncycastle.asn1.ASN1Sequence;
@@ -52,6 +51,8 @@ import java.util.Iterator;
 import java.util.UUID;
 
 import static org.apache.hadoop.hdds.HddsConfigKeys.OZONE_METADATA_DIRS;
+import static org.apache.hadoop.hdds.security.x509.certificate.utils.CertificateSignRequest.getDistinguishedNameFormat;
+import static org.apache.hadoop.hdds.security.x509.certificate.utils.CertificateSignRequest.getPkcs9Extensions;
 
 /**
  * Certificate Signing Request.
@@ -88,7 +89,7 @@ public class TestCertificateSignRequest {
     PKCS10CertificationRequest csr = builder.build();
 
     // Check the Subject Name is in the expected format.
-    String dnName = String.format(SecurityUtil.getDistinguishedNameFormat(),
+    String dnName = String.format(getDistinguishedNameFormat(),
         subject, scmID, clusterID);
     Assertions.assertEquals(dnName, csr.getSubject().toString());
 
@@ -101,7 +102,7 @@ public class TestCertificateSignRequest {
 
     // Verify CSR with attribute for extensions
     Assertions.assertEquals(1, csr.getAttributes().length);
-    Extensions extensions = SecurityUtil.getPkcs9Extensions(csr);
+    Extensions extensions = getPkcs9Extensions(csr);
 
     // Verify key usage extension
     Extension keyUsageExt = extensions.getExtension(Extension.keyUsage);
@@ -148,7 +149,7 @@ public class TestCertificateSignRequest {
     PKCS10CertificationRequest csr = builder.build();
 
     // Check the Subject Name is in the expected format.
-    String dnName = String.format(SecurityUtil.getDistinguishedNameFormat(),
+    String dnName = String.format(getDistinguishedNameFormat(),
         subject, scmID, clusterID);
     Assertions.assertEquals(dnName, csr.getSubject().toString());
 
@@ -161,7 +162,7 @@ public class TestCertificateSignRequest {
 
     // Verify CSR with attribute for extensions
     Assertions.assertEquals(1, csr.getAttributes().length);
-    Extensions extensions = SecurityUtil.getPkcs9Extensions(csr);
+    Extensions extensions = getPkcs9Extensions(csr);
 
     // Verify key usage extension
     Extension sanExt = extensions.getExtension(Extension.keyUsage);
@@ -230,7 +231,7 @@ public class TestCertificateSignRequest {
     PKCS10CertificationRequest csr = builder.build();
 
     // Check the Subject Name is in the expected format.
-    String dnName = String.format(SecurityUtil.getDistinguishedNameFormat(),
+    String dnName = String.format(getDistinguishedNameFormat(),
         subject, scmID, clusterID);
     Assertions.assertEquals(dnName, csr.getSubject().toString());
